in main_planetary_motion.py:  Improved variable names to match what is used in the function call
in twoBodyProblemAreaSwept.py: Fixed typo in description at the top, added a bit to clarify meaning as well. Removed unused variables. Changed 'gd = graph()' to 'graphSetup = graph()' to make it more clear what that line does, as 'gd' is meaningless, even to me.

Moved nBodyExperimental.py to its own repository, because it isn't relevant to this one.