Added the recordTelemetry() and handlePeriodCounting() object methods to planetObjectGenerator.py, and prevented these from breaking things if they are called from the sun.
Switched the method of calculating distance between various objects in the static portion of the model in trueAndNotTrueThreeBodyComparison.py to better match the way it is done in trueThreeBodyProblem.py (which is superior in every measurable way). Also changed color of some planet objects after their creation to make it easier to keep track of which one is which.